Change color to prevent visibility issue
Caption seems to appear in lower area of image on some screens so that the brown
font blends into the brown background.

I changed it to a lighter font when the image is shown so that the caption  should appear well regardless of where it
rests in the image.

I then changed the coloring in hover to make it flow more naturally.

I also removed extra px designation and space.